backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.3.0

# Database
motor==3.3.2
pymongo==4.6.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.1

# Line Bot SDK
line-bot-sdk==3.6.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1

# Date & Time
python-dateutil==2.8.2

# CSV & Data Export
pandas==2.1.4
openpyxl==3.1.2

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1

